Add a blind auction example
The `BlindAuction` example contract supports a blind auction with the
following properties:
 * bids remain hidden
 * winning bid remains hidden
 * winner address becomes public at the end

In order to support hidden bids, we link the auction to the
`EncryptedERC20` contract, such that bids are in this token. We utilize
the allowances mechanism to approve use of user tokens on their behalf
by the auction contract.

The flow is:
1. Auction owner creates the auction contract.
2. Users approve use of tokens in the ERC20 contract via `approve()`.
3. Users call `bid()` to bid an amount less than or equal to the
approved token amount.
4. Auction owner stops the auction.
5. Users can call `doIHaveHighestBid() to know if they have the highest
bid.
6. If user has highest bid, they can call `claim()` to claim the object.
At that point, we decrypt a boolean that shows whether the caller has
the highest bid. No other decryptions are done.
7. If user doesn't have the highest bid, they can `withdraw()` their
bid.

Currently, we emit an event to denote that the winner has claimed the
"object" of the auction.

If two or more bidders bid the same value, whoever claims first is the
winner. All other bidders can only withdraw their bids at that point.